Alterações e Adições no Backend

+++ backend/src/main/java/com/conectin/conectin/repository/UsuarioRepository.java
@@ -6,8 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.conectin.conectin.entities.Usuario;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     Optional<Usuario> findByEmail(String email);
-
     boolean existsByEmail(String email);
 }
